autoML_scripts/productive/requirements.txt
pandas
mljar
mljar-supervised
boto3
minio
python-dotenv
flask